DPC-499: Rename the V2 aggregation engine
